% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblast.R
\name{flycircuit_nblast}
\alias{flycircuit_nblast}
\title{Fetch precomputed nblast scores between flycircuit neurons}
\usage{
flycircuit_nblast(query, target, normalisation = c("normalised", "mean",
  "raw"))
}
\arguments{
\item{query}{Vectors of FlyCircuit identifiers (Neuron name, gene_name
or integer idid are all acceptable since they will be processed by
\code{\link{fc_gene_name}}).}

\item{target}{Vectors of FlyCircuit identifiers (Neuron name, gene_name
or integer idid are all acceptable since they will be processed by
\code{\link{fc_gene_name}}).}

\item{normalisation}{the type of normalisation procedure that should be
carried out, selected from \code{'raw'}, \code{'normalised'} or
\code{'mean'} (i.e. the average of normalised scores in both directions).}
}
\description{
Fetch precomputed nblast scores between flycircuit neurons
}

