% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblast.R
\name{flycircuit_topn}
\alias{flycircuit_topn}
\title{NBLAST a single flycircuit neuron against database and return scores}
\usage{
flycircuit_topn(query, n = 50, ...)
}
\arguments{
\item{query}{A flycircuit identifier}

\item{n}{The number of scores to return}
}
\value{
a data.frame in descending score order with columns \itemize{

  \item id Flycircuit Neuron name

  \item score NBLAST score

  }
}
\description{
NBLAST a single flycircuit neuron against database and return scores
}
\examples{
\dontrun{
## example of remote call via curl
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn -d 'query="FruMARCM-M002262_seg001"'
# (return value looks like this and includes sesion key)
/ocpu/tmp/x05f62b2975/R/.val
/ocpu/tmp/x05f62b2975/messages
/ocpu/tmp/x05f62b2975/stdout
/ocpu/tmp/x05f62b2975/source
/ocpu/tmp/x05f62b2975/console
/ocpu/tmp/x05f62b2975/info
/ocpu/tmp/x05f62b2975/files/DESCRIPTION

# now collect results
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/json
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/csv

# specify number of hits, use "neuron" identifier
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn -d 'query="fru-M-200266"&n=10'
}
}

