% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flynblastscores_storage_dir}
\alias{flynblastscores_storage_dir}
\title{Return the location in which external data (i.e. score matrices) are stored}
\usage{
flynblastscores_storage_dir()
}
\value{
Character vector defining path
}
\description{
Return the location in which external data (i.e. score matrices) are stored
}
