% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nblast.R
\name{flycircuit_topn}
\alias{flycircuit_gmr_topn}
\alias{flycircuit_topn}
\title{NBLAST a single flycircuit neuron against database and return scores}
\usage{
flycircuit_topn(query, n = 50, ...)

flycircuit_gmr_topn(query, n = 50)
}
\arguments{
\item{query}{A flycircuit identifier}

\item{n}{The number of scores to return}

\item{...}{Additional arguments passed to \code{\link{flycircuit_nblast}}}
}
\value{
a data.frame in descending score order with columns \itemize{

  \item id Flycircuit Neuron name (or GMR id for \code{flycircuit_gmr_topn})

  \item score NBLAST score (normalised -1 (bad) to +1 (perfect)). Scores
  above 0 are potentially interesting, scores above about 0.4 seriously
  interesing.

  }
}
\description{
NBLAST a single flycircuit neuron against database and return scores

\code{flycircuit_gmr_topn} finds the top GMR Gal4 lines matching
  a FlyCircuit neuron query.
}
\examples{
\dontrun{
## example of remote call via curl
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn -d 'query="FruMARCM-M002262_seg001"'
# (return value looks like this and includes sesion key)
/ocpu/tmp/x05f62b2975/R/.val
/ocpu/tmp/x05f62b2975/messages
/ocpu/tmp/x05f62b2975/stdout
/ocpu/tmp/x05f62b2975/source
/ocpu/tmp/x05f62b2975/console
/ocpu/tmp/x05f62b2975/info
/ocpu/tmp/x05f62b2975/files/DESCRIPTION

# now collect results
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/json
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/csv

# specify number of hits, use "neuron" identifier
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn -d 'query="fru-M-200266"&n=10'
}
flycircuit_gmr_topn("VGlut-F-200269")
}

