% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flynblastscores-package.R
\docType{package}
\name{flynblastscores-package}
\alias{flynblastscores-package}
\alias{flynblastscores}
\title{Functions to return pre-computed NBLAST scores}
\description{
Functions to return pre-computed NBLAST scores
}
\keyword{package}
